requirements.txt
Django==1.11
lxml==4.1.1
PyHyphen==3.0.1
psycopg2==2.7.3.2
djangorestframework==3.7.7
requests==2.18.4
python-magic==0.4.15
html-sanitizer==1.3.0